split server request interface
because it's unlikely the same consumer would need to create a server-request from globals, as well as an empty request - they're likely two different operations; typically, creating from superglobals is something only a dispatcher or front-controller would need to do. I think that we should separate these operations, which also is more consistent with other interfaces dealing with only a single responsibility.